+++ test/e2e/test_scheduler.go
@@ -25,8 +25,8 @@ import (
 	"knative.dev/pkg/test/helpers"
 
 	schemasv1 "github.com/google/knative-gcp/pkg/schemas/v1"
-	"github.com/google/knative-gcp/test/e2e/lib"
-	"github.com/google/knative-gcp/test/e2e/lib/resources"
+	"github.com/google/knative-gcp/test/lib"
+	"github.com/google/knative-gcp/test/lib/resources"
 )
 
 // SmokeCloudSchedulerSourceTestHelper tests if a CloudSchedulerSource object can be created to ready state.
