+++ test/lib/deletion.go
