+++ test/lib/helpers/broker_event_transformation_test_with_source.go
@@ -37,9 +37,9 @@ import (
 	_ "k8s.io/client-go/plugin/pkg/client/auth/gcp"
 
 	schemasv1 "github.com/google/knative-gcp/pkg/schemas/v1"
-	"github.com/google/knative-gcp/test/e2e/lib"
-	"github.com/google/knative-gcp/test/e2e/lib/metrics"
-	"github.com/google/knative-gcp/test/e2e/lib/resources"
+	"github.com/google/knative-gcp/test/lib"
+	"github.com/google/knative-gcp/test/lib/metrics"
+	"github.com/google/knative-gcp/test/lib/resources"
 )
 
 /*
