+++ test/lib/resources/knative_gcp.go
