+++ test/lib/secrets.go
