+++ test/lib/typemetas.go
@@ -19,7 +19,7 @@ package lib
 import (
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 
-	"github.com/google/knative-gcp/test/e2e/lib/resources"
+	"github.com/google/knative-gcp/test/lib/resources"
 )
 
 var JobTypeMeta = batchTypeMeta(resources.JobKind)
